pyproject.toml
[project]
name = "rag-llmops-demo-apps"
version = "0.1.0"
description = "Corporate Affairs RAG Chatbot on Databricks Apps (migrated from Model Serving)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "mlflow>=3.8.0rc0",
    "databricks-sdk",
    "databricks-agents",
    "databricks-vectorsearch",
    "databricks-openai",
    "openai>=1.0",
    "python-dotenv",
]

[tool.uv]
prerelease = "allow"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent_server", "scripts"]

[dependency-groups]
dev = [
    "hatchling>=1.27.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
quickstart = "scripts.quickstart:main"
start-app = "scripts.start_app:main"
start-server = "agent_server.start_server:main"
agent-evaluate = "agent_server.evaluate_agent:evaluate"
